Move AbsenceMapping into calendarintegration package
