+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/AbsenceMappingRepository.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.absence;
+package org.synyx.urlaubsverwaltung.calendarintegration;
 
 import org.springframework.data.repository.CrudRepository;
