+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/AbsenceMappingServiceImpl.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.absence;
+package org.synyx.urlaubsverwaltung.calendarintegration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
