+++ src/main/java/org/synyx/urlaubsverwaltung/calendarintegration/AbsenceMappingType.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.absence;
+package org.synyx.urlaubsverwaltung.calendarintegration;
 
 public enum AbsenceMappingType {
